backend/requirements.txt
# Django LTS and grappelli
django==1.8
django-grappelli==2.7.1

djangorestframework==3.3.2
# Markdown support for the browsable API.
markdown==2.3.1
django-filter==0.12.0

# http://djangothumbnails.com/
django-thumbs==0.4

# http://django-extensions-zh.readthedocs.io/
django-extensions==1.6.7

requests==2.10.0

pillow

redis==2.10.5
django-redis==4.2.0
MySQL-python==1.2.5

uWSGI==2.0.13
